feat(ai): add vitest-free axiom/ai/evals/scorers entry point
The axiom/ai/evals entry point unconditionally imports vitest at the
top level via the Eval class. This causes production deployments on
Vercel to fail with ERR_MODULE_NOT_FOUND because vitest (a
devDependency) is unavailable at runtime.

Split scorer type definitions into scorer.types.ts and create a clean
barrel in scorers.ts that exports createScorer as Scorer without
pulling in vitest. This gives production code a vitest-free import
path while keeping axiom/ai/evals unchanged for offline eval users.
