+++ examples/online-evals-comprehensive/online-evals-comprehensive.ts
@@ -21,7 +21,8 @@
 import { generateObject, generateText } from 'ai';
 import { createOpenAI } from '@ai-sdk/openai';
 import { type SpanContext } from '@opentelemetry/api';
-import { withSpan, wrapAISDKModel, Scorer, onlineEval } from 'axiom/ai';
+import { withSpan, wrapAISDKModel, onlineEval } from 'axiom/ai';
+import { Scorer } from 'axiom/ai/evals/scorers';
 import { z } from 'zod';
 import { initializeTelemetry, flushTelemetry } from './instrumentation';
